Create components for splitting the youtube page into parts so its easier to handle
create a searchbar class that sets the state to an empty string then sets the value of the user input to the input. Export so that we can call in index

create a videolist function that takes in videolistitems, and maps through them to create an array of video items with unique keys (etag from API)  and then passes that array into a ul. Export

Create a videolistitem function that creates list items to pass to the videolist. Export

create a videodetail file