import additional classes and functions
Get our api key for youtube, store as a variable for future use.

change App from a function into a class that extends component, with a constructor that sets the state equal to an empty array, then uses the API call to fill that array with the results.

renders a div that contains the searchbar and videolist components.

renders the App class to the DOM